basic-connection/pyproject.toml
[project]
name = "lm-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
